requirements.txt
requests
streamlit==1.38.0
pandas==2.2.2
gTTS==2.5.3
PyPDF2==3.0.1
google-generativeai==0.7.2
openai==1.50.2
streamlit-mic-recorder==0.0.5
openpyxl==3.1.5